Improve regex matching, allowing trailing |
The regex now supports trailing | in favor of dropping support to
seperate multiple vendors by __
The __ is only to be used as the delimiter between the whole vendor
prefix and the desired manifest key. __ will end the vendor prefix
matching. Everything after it (if a valid vendor or otherwise) will
be interpreted as part of the desired manifest key.

Typos in vendors will result in a complete missmatch of the vendor
prefix detection; The whole key, as is, will be part of every generated
manifest (This will throw errors in the browser upon adding).

Examples:
	Valid:
	__chrome__key: value       -> key: value
	__chrome|edge__key: value  -> key: value

	Trailing |:
	__chrome|__key: value      -> key: value
	__chrome|edge|__key: value -> key: value

	__ between vendors:
	__chrome__edge__key: value -> edge__key: value
	__chrome____key: value     -> __key: value

	Typo in vendor:
	__chrome|opea__key: value  -> __chrome|opea__key: value
	__chrome__opea__key: value -> opea__key: value
	__opea__chrome__key: value -> __opea__chrome__key: value
	__chrome__opea__firefox__key: value -> opea__firefox__key: value